inference-setup/sample-submission/MT2024131/requirements.txt
torch>=2.0.0
transformers>=4.41.1  # Important: version needs to be at least 4.41.1 for proper processor support
bitsandbytes>=0.40.0
accelerate>=0.20.0
peft>=0.5.0
pillow>=9.0.0
pandas>=1.3.0
tqdm>=4.65.0  # Added for progress bars
flash-attn